# api/docs/swagger.yaml
openapi: 3.0.3
info:
  title: StudentApp API
  description: API для студенческого приложения с геймификацией
  version: 2.0.0
  contact:
    name: Support
    email: support@studentapp.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.studentapp.com/api
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        login:
          type: string
          example: "23-ПМбо-014"
        name:
          type: string
          example: "Иванов Иван Иванович"
        role:
          type: string
          enum: [admin, teacher, staff, student]
          example: "student"
        group:
          type: string
          example: "23-ПМбо-1"
        course:
          type: integer
          example: 3
        faculty:
          type: string
          example: "ИВИТШ"
        email:
          type: string
          example: "ivan@student.ru"
        phone:
          type: string
          example: "+7 (999) 123-45-67"
        total_xp:
          type: integer
          example: 1250
        coins:
          type: integer
          example: 500
    LoginRequest:
      type: object
      required:
        - login
        - password
      properties:
        login:
          type: string
          example: "23-ПМбо-014"
        password:
          type: string
          example: "student123"
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIs..."
        user:
          $ref: '#/components/schemas/User'

    # ========== РАСПИСАНИЕ ==========
    Schedule:
      type: object
      properties:
        id:
          type: integer
        group_name:
          type: string
        discipline:
          type: string
        teacher:
          type: string
        start_time:
          type: string
          example: "11:50"
        end_time:
          type: string
          example: "13:20"
        date:
          type: string
          format: date
        day_of_week:
          type: integer
        audience:
          type: string
        lesson_type:
          type: string
          enum: [lecture, lab, practice]
        comment:
          type: string

    # ========== АЧИВКИ ==========
    Achievement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
          example: "🌙 Полуночник"
        description:
          type: string
          example: "Сдать домашнее задание после 2 часов ночи"
        xp:
          type: integer
          example: 50
        icon:
          type: string
          example: "🌙"
        rarity:
          type: string
          enum: [common, rare, epic, legendary]
        category:
          type: string
        is_unlocked:
          type: boolean

    # ========== НОВОСТИ ==========
    News:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        short_text:
          type: string
        full_text:
          type: string
        image_url:
          type: string
        category:
          type: string
        date:
          type: string
          format: date
        is_pinned:
          type: boolean
        views:
          type: integer

    # ========== МЕРОПРИЯТИЯ ==========
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        short_text:
          type: string
        full_text:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
        type:
          type: string
          enum: [sport, school, student]
        category:
          type: string
        location:
          type: string
        price:
          type: string
        organizer:
          type: string
        available_spots:
          type: integer
        is_registered:
          type: boolean

    # ========== ДЕДЛАЙНЫ ==========
...
    Chat:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [private, group, course]
        last_message:
          type: string
        unread_count:
          type: integer
    Message:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        user_name:
          type: string
        text:
          type: string
        is_mine:
          type: boolean
        created_at:
          type: string
          format: date-time

    # ========== КОИНЫ ==========
    Balance:
      type: object
      properties:
        balance:
          type: integer
    ShopItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: integer
        icon:
          type: string

    # ========== ОШИБКИ ==========
...
security:
  - bearerAuth: []
...
paths:
...
  /login:
    post:
      tags: [Auth]
      summary: Вход в систему
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Неверный логин или пароль
  /register:
    post:
      tags: [Auth]
      summary: Регистрация нового пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - login
                - password
                - name
              properties:
                login:
                  type: string
                password:
                  type: string
                name:
                  type: string
                role:
                  type: string
                group:
                  type: string
      responses:
        '201':
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'

  # ========== USERS ==========
  /users/me:
    get:
      tags: [Users]
      summary: Получить свой профиль
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      tags: [Users]
      summary: Обновить свой профиль
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                phone:
                  type: string
                telegram:
                  type: string
      responses:
        '200':
          description: Профиль обновлен

  # ========== SCHEDULE ==========
  /schedule:
    get:
      tags: [Schedule]
      summary: Получить расписание
      security:
        - bearerAuth: []
      parameters:
        - name: group_name
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Список занятий
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schedule'

    post:
      tags: [Schedule]
      summary: Добавить занятие (админ/учитель)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schedule'
      responses:
        '201':
          description: Занятие добавлено

  # ========== ACHIEVEMENTS ==========
  /achievements:
    get:
      tags: [Achievements]
      summary: Получить все ачивки
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список ачивок
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Achievement'
  /achievements/me:
    get:
      tags: [Achievements]
      summary: Получить свои ачивки
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Мои ачивки
  /achievements/me/xp:
    get:
      tags: [Achievements]
      summary: Получить общее количество XP
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Количество XP
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_xp:
                    type: integer
...
  /news:
    get:
      tags: [News]
      summary: Получить новости
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Список новостей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/News'
  /news/pinned:
    get:
      tags: [News]
      summary: Получить закрепленные новости
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Закрепленные новости

  # ========== EVENTS ==========
  /events:
    get:
      tags: [Events]
      summary: Получить мероприятия
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [sport, school, student]
      responses:
        '200':
          description: Список мероприятий
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
  /events/{id}/register:
    post:
      tags: [Events]
      summary: Записаться на мероприятие
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Запись успешна

    delete:
      tags: [Events]
      summary: Отменить запись на мероприятие
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Запись отменена

  # ========== SKILL TREE ==========
...
  /games:
    get:
      tags: [Games]
      summary: Получить список мини-игр
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список игр
  /games/play:
    post:
      tags: [Games]
      summary: Отправить результат игры
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                game_id:
                  type: integer
                score:
                  type: integer
      responses:
        '200':
          description: Результат сохранен
  /games/{id}/leaderboard:
    get:
      tags: [Games]
      summary: Таблица лидеров игры
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Таблица лидеров

  # ========== PLANNER ==========
...
  /chats:
    get:
      tags: [Chats]
      summary: Получить список чатов
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список чатов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Chat'
  /chats/{id}/messages:
    get:
      tags: [Chats]
      summary: Получить сообщения чата
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Сообщения чата
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'

    post:
      tags: [Chats]
      summary: Отправить сообщение
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                file_url:
                  type: string
      responses:
        '201':
          description: Сообщение отправлено
  /chats/private/{userId}:
    post:
      tags: [Chats]
      summary: Создать приватный чат
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Чат создан
  /chats/group:
    post:
      tags: [Chats]
      summary: Создать групповой чат
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                members:
                  type: array
                  items:
                    type: integer
      responses:
        '201':
          description: Чат создан

  # ========== NOTIFICATIONS ==========
  /notifications:
    get:
      tags: [Notifications]
      summary: Получить уведомления
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Список уведомлений
  /notifications/unread:
    get:
      tags: [Notifications]
      summary: Количество непрочитанных уведомлений
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Количество
          content:
            application/json:
              schema:
                type: object
                properties:
                  unread_count:
                    type: integer
  /notifications/{id}/read:
    put:
      tags: [Notifications]
      summary: Отметить уведомление прочитанным
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Отмечено
  /notifications/read-all:
    put:
      tags: [Notifications]
      summary: Отметить все уведомления прочитанными
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Все отмечены

  # ========== COINS ==========
  /coins/balance:
    get:
      tags: [Coins]
      summary: Получить баланс коинов
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Баланс
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Balance'
  /coins/transactions:
    get:
      tags: [Coins]
      summary: История транзакций
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Список транзакций
  /coins/shop:
    get:
      tags: [Coins]
      summary: Товары в магазине
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список товаров
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShopItem'
  /coins/shop/{itemId}/buy:
    post:
      tags: [Coins]
      summary: Купить товар
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Товар куплен

  # ========== WEBSOCKET ==========
  /ws:
    get:
      tags: [WebSocket]
      summary: WebSocket соединение для чатов
      security:
        - bearerAuth: []
      responses:
        '101':
          description: Switching Protocols